Fw19 0529 day5

+++ TravelZilla/src/main/java/com/travelzilla/controllers/CustomerController.java
@@ -16,10 +16,15 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.travelzilla.exceptions.AdminException;
 import com.travelzilla.exceptions.CustomerException;
+import com.travelzilla.exceptions.SessionException;
 import com.travelzilla.models.Customer;
+import com.travelzilla.models.Session;
+import com.travelzilla.models.UserType;
 import com.travelzilla.services.CustomerServices;
 import com.travelzilla.services.SessionServices;
 
@@ -42,32 +47,44 @@ public class CustomerController {
 	}
 	
 	@DeleteMapping("/deletecustomer/{cid}")
-	public ResponseEntity<Customer> deleteCustomerbyIdHandler(@PathVariable Integer customerId) throws CustomerException{
-		Customer dc = customerServices.deleteCustomer(customerId) ;
-		return new ResponseEntity<Customer>(dc, HttpStatus.OK);
+	public ResponseEntity<Customer> deleteCustomerbyIdHandler(@PathVariable Integer customerId, @RequestParam("sessionKey")String sessionKey) throws CustomerException, SessionException{
+		Session session = sServices.getASessionByKey(sessionKey);
+		if(session.getUserId()==customerId && session.getUserType()==UserType.CUSTOMER) {
+			Customer dc = customerServices.deleteCustomer(customerId) ;
+			return new ResponseEntity<Customer>(dc, HttpStatus.OK);
+		}
+		throw new CustomerException("Please login with the correct credentials");
 	}
 	
 	@PutMapping("/updatecustomer")
-	public ResponseEntity<Customer> updateCustomerHandler(@RequestBody Customer customer) throws CustomerException{
+	public ResponseEntity<Customer> updateCustomerHandler(@RequestBody Customer customer, @RequestParam("sessionKey")String sessionKey) throws CustomerException, SessionException{
+		Session session = sServices.getASessionByKey(sessionKey);
+		if(session.getUserId()==customer.getCustomerId() && session.getUserType()==UserType.CUSTOMER) {
+			Customer updatedCustomer= customerServices.updateCustomer(customer);
 		
-		Customer updatedCustomer= customerServices.updateCustomer(customer, null);
-		
-		return new ResponseEntity<Customer>(updatedCustomer,HttpStatus.ACCEPTED);
-		
+			return new ResponseEntity<Customer>(updatedCustomer,HttpStatus.ACCEPTED);
+		}
+		throw new CustomerException("Please login with the correct credentials");
 	}
 	
 	@GetMapping("/viewcustomer/{id}")
-	public ResponseEntity<Customer> viewCustomer(@PathVariable("id") Integer id) throws CustomerException{
+	public ResponseEntity<Customer> viewCustomer(@PathVariable("id") Integer id, @RequestParam("sessionKey")String sessionKey) throws CustomerException, SessionException{
+		Session session = sServices.getASessionByKey(sessionKey);
+		if(session.getUserId()== id && session.getUserType()==UserType.CUSTOMER) {
+			return new ResponseEntity<>(customerServices.viewCustomer(id), HttpStatus.OK);
+		}
 		
-		return new ResponseEntity<>(customerServices.viewCustomer(id), HttpStatus.OK);
+		throw new CustomerException("Please login with the correct credentials");
 		
 	}
 	
 	@GetMapping("/viewallcustomers")
-	public ResponseEntity<List<Customer>> viewAllCustomers() throws CustomerException{
-		
-		return new ResponseEntity<List<Customer>>(customerServices.viewAllCustomer(),HttpStatus.OK);
-		
+	public ResponseEntity<List<Customer>> viewAllCustomers(@RequestParam("sessionKey")String sessionKey) throws CustomerException, SessionException{
+		Session session = sServices.getASessionByKey(sessionKey);
+		if(session.getUserType() == UserType.ADMIN) {
+			return new ResponseEntity<List<Customer>>(customerServices.viewAllCustomer(),HttpStatus.OK);
+		}
+		throw new CustomerException("Access Denied! Only admin can view all customers.");
 	}
